feat: add a glossary generator for all function modules
